Agentic/marketing-swarm/backend/requirements.txt
# Core Framework
# Note: CrewAI is not compatible with Python 3.13, using alternative approach
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-socketio==5.11.0
python-multipart==0.0.6
websockets==12.0

# AI/ML
openai==1.12.0
# Removed langchain and crewai due to Python 3.13 compatibility issues

# Database
sqlalchemy[asyncio]==2.0.25
aiosqlite==0.19.0
asyncpg==0.29.0          # Async PostgreSQL driver for Cloud SQL
psycopg2-binary==2.9.9   # Sync PostgreSQL driver (for migrations)
greenlet==3.0.3          # Required for SQLAlchemy async
alembic==1.13.1

# Caching (optional)
# redis==5.0.1
# aioredis==2.0.1

# Monitoring & Safety
# psutil==5.9.8  # Optional, can be installed separately if needed
prometheus-client==0.19.0
# sentry-sdk==1.40.0  # Optional for production

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cryptography==42.0.2

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
tenacity==8.2.3
# asyncio is built-in, no need to install

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Logging
loguru==0.7.2
python-json-logger==2.0.7

# CORS support
fastapi-cors==0.0.6